+++ src/main.rs
@@ -4,22 +4,18 @@ extern crate regex;
 extern crate reqwest;
 
 mod actor;
-mod chat_client;
-mod claude_client;
 mod cmd;
 mod config;
 mod docker;
 mod emoji;
 mod env;
-mod local_client;
 mod logging;
 mod persistence;
 mod shutdown;
 mod web;
 
-use chat_client::create_chat_client;
-use clap::{Parser, ValueEnum};
-use cmd::Handler;
+use clap::Parser;
+use cmd::{chat_mode::LocalClient, Handler};
 use config::Config;
 use env::Environment;
 use persistence::PersistentStore;
@@ -50,12 +46,6 @@ impl std::str::FromStr for WebBindAddress {
   }
 }
 
-#[derive(ValueEnum, Clone, Debug)]
-pub enum ChatClientType {
-  Claude,
-  Local,
-}
-
 #[derive(Parser)]
 #[command(name = "disbot")]
 #[command(about = "Discord bot with admin web interface")]
@@ -75,10 +65,6 @@ struct Cli {
   /// Web server bind address (IP address, "lan" for LAN IP, or "0.0.0.0" for all interfaces)
   #[arg(long, default_value = "0.0.0.0")]
   web_bind_address: WebBindAddress,
-
-  /// Chat client backend to use
-  #[arg(long, value_enum, default_value = "claude")]
-  chat_client: ChatClientType,
 }
 
 pub struct HttpClient;
@@ -114,15 +100,8 @@ async fn main() -> Result<(), anyhow::Error> {
   // Persistence restoration happens in the ready event handler where actor handles are available
   let persistence = Arc::new(PersistentStore::new(&config.db_path)?);
 
-  // Create chat client based on CLI flag
-  let chat_client = match create_chat_client(cli.chat_client, &config, persistence.clone()).await {
-    Ok(client) => client,
-    Err(e) => {
-      error!("Failed to create chat client: {}", e);
-      std::process::exit(1);
-    }
-  };
-
+  // Create local chat client
+  let chat_client = LocalClient::new(&config, persistence.clone()).await?;
   let emoji = emoji::EmojiLookup::new(&config);
   let http = reqwest::Client::new();
